use CacheService for all requests #1
