+++ src/main/java/de/hpi/cache/api/CacheController.java
@@ -1,18 +1,13 @@
 package de.hpi.cache.api;
 
 import de.hpi.cache.dto.IdealoOffer;
-import de.hpi.cache.persistence.ShopOffer;
 import de.hpi.cache.persistence.repositories.ShopOfferRepositoryImpl;
 import de.hpi.cache.services.CacheService;
-import de.hpi.cache.services.IdealoBridge;
-import de.hpi.cache.services.WarmupService;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
-
-import static org.springframework.web.bind.annotation.RequestMethod.GET;
 
 @RestController
 @Slf4j
@@ -20,26 +15,25 @@ import static org.springframework.web.bind.annotation.RequestMethod.GET;
 @RequiredArgsConstructor
 public class CacheController {
     private final CacheService cacheService;
-    private final WarmupService warmupService;
     private final ShopOfferRepositoryImpl shopOfferRepository;
 
     @RequestMapping(value = "/getOffer/{shopID}", method = RequestMethod.GET, produces = "application/json")
     public IdealoOffer getOffer(@PathVariable long shopID, @RequestParam(value = "phase") short phase){
-        return getShopOfferRepository().getOffer(shopID, phase);
+        return getCacheService().getOffer(shopID, phase);
     }
 
-    @RequestMapping(value = "deleteOffer/{shopID}", method = RequestMethod.GET)
+    @RequestMapping(value = "/deleteOffer/{shopID}", method = RequestMethod.DELETE)
     public void deleteOffer(@PathVariable long shopID, @RequestParam(value = "offerKey") String offerKey){
-        getShopOfferRepository().deleteOffer(shopID, offerKey);
+        getCacheService().deleteOffer(shopID, offerKey);
     }
 
-    @RequestMapping(value = "deleteAll/{shopID}", method = RequestMethod.GET)
+    @RequestMapping(value = "/deleteAll/{shopID}", method = RequestMethod.DELETE)
     public void deleteAll(@PathVariable long shopID){
-        getShopOfferRepository().deleteAll(shopID);
+        getCacheService().deleteAll(shopID);
     }
 
-    @RequestMapping(value = "/warmup/{shopID}", method = GET)
+    @RequestMapping(value = "/warmup/{shopID}", method = RequestMethod.GET)
     public void warmup(@PathVariable long shopID){
-        getWarmupService().warmup(shopID);
+        getCacheService().warmup(shopID);
     }
 }
